### 2023-12-09 #6131 @editorialbot
[PRE REVIEW]: REDCapTidieR: Extracting complex REDCap databases into tidy tables
**Submitting author:** @rsh52 (<a href="http://orcid.org/0009-0005-6496-8154">Richard Hanna</a>)
**Repository:** https://github.com/CHOP-CGTInformatics/REDCapTidieR
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
