### 2023-12-09 #6131 @arfon
@rsh52 – thanks for your submission to JOSS. I think you need to change the last affiliation from this:

```
- name: Department of Pediatrics, The University of Oklahoma Health Sciences Center, College of Medicine, Oklahoma City, Oklahoma, USA
- index: 7
```

to this:

```
- name: Department of Pediatrics, The University of Oklahoma Health Sciences Center, College of Medicine, Oklahoma City, Oklahoma, USA
  index: 7
```

We're currently managing a large backlog of submissions and the editor most appropriate for your area is already rather busy.
