### 2023-12-11 0eb5790
Address JOSS feedback :pencil:
Fix affiliation index, fix DOIs in bib
