+++ paper/paper.bib
@@ -5,7 +5,7 @@
  pages = {103208},
  year = {2019},
  issn = {1532-0464},
- doi = {https://doi.org/10.1016/j.jbi.2019.103208},
+ doi = {10.1016/j.jbi.2019.103208},
  url = {https://www.sciencedirect.com/science/article/pii/S1532046419301261},
  author = {Paul A. Harris and Robert Taylor and Brenda L. Minor and Veida Elliott and Michelle Fernandez and Lindsay O'Neal and Laura McLeod and Giovanni Delacqua and Francesco Delacqua and Jacqueline Kirby and Stephany N. Duda},
  keywords = {Medical informatics, Electronic data capture, Clinical research, Translational research},
@@ -20,7 +20,7 @@
  pages = {377-381},
  year = {2009},
  issn = {1532-0464},
- doi = {https://doi.org/10.1016/j.jbi.2008.08.010},
+ doi = {10.1016/j.jbi.2008.08.010},
  url = {https://www.sciencedirect.com/science/article/pii/S1532046408001226},
  author = {Paul A. Harris and Robert Taylor and Robert Thielke and Jonathon Payne and Nathaniel Gonzalez and Jose G. Conde},
  keywords = {Medical informatics, Electronic data capture, Clinical research, Translational research},
