+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   diving_deeper: diving_deeper.html
   export_to_xlsx: export_to_xlsx.html
   glossary: glossary.html
-last_built: 2023-12-05T18:15Z
+last_built: 2023-12-11T14:22Z
 urls:
   reference: https://chop-cgtinformatics.github.io/REDCapTidieR/reference
   article: https://chop-cgtinformatics.github.io/REDCapTidieR/articles
