### 2023-12-11 #6131 @rsh52
@arfon Thank you for the feedback and no problem about the waitlist, we understand.

I updated the affiliation, thank you for catching that, as well as the invalid DOIs the bot picked up. Attempting paper regeneration.
