### 2023-12-12 #6131 @arfon
☝️ the commands to @editorialbot have to be at the start of a message.
