### 2024-01-20 #6131 @arfon
:wave: @mikemahoney218 – would you be willing to edit this submission for JOSS? / cc @martinfleis for visibility
