### 2024-01-20 #6131 @mikemahoney218
Happy to take this one on!

@rsh52 , would you be able to suggest 5 reviewers for this package? Reviewers can come from either [our reviewer database](https://reviewers.joss.theoj.org/) or your own network, so long as they don't have a [conflict of interest](https://joss.theoj.org/about#ethics). After that, I should be able to start pinging people for a review next week!
