### 2024-01-22 #6131 @rsh52
> Happy to take this one on!
>
> @rsh52 , would you be able to suggest 5 reviewers for this package? Reviewers can come from either [our reviewer database](https://reviewers.joss.theoj.org/) or your own network, so long as they don't have a [conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). After that, I should be able to start pinging people for a review next week!

Hi @mikemahoney218, thank you for working with us! Can we suggest the following 5 reviewers from your database:

- [simonpcouch](https://reviewers.joss.theoj.org/reviewers/2171)
- [RhysPeploe](https://reviewers.joss.theoj.org/reviewers/3578)
- [JerryChiaRuiChang](https://reviewers.joss.theoj.org/reviewers/3367)
- [jackmwolf](https://reviewers.joss.theoj.org/reviewers/2916)
- [brunomontezano](https://reviewers.joss.theoj.org/reviewers/2055)

Thank you! Let me know what else you need from us.
