### 2024-01-24 #6277 @mikemahoney218
👋🏼 @rsh52 , @RhysPeploe , @spgarbet , this is the review thread for {REDCapTidieR}. Just about all of our communications will happen here from now on :smile: .

We might have one more reviewer join us, but I figured it made sense to get things started while I wait to hear back there :smile:

```@editorialbot generate my checklist```

as the top of a new comment in this thread. For best results, don't include anything else in the comment!

This will create a checklist that walks through the JOSS submission requirements. As you go over the submission, please check any items that you feel have been satisfied. The first comment in this thread also contains links to the JOSS reviewer guidelines.

We aim for reviews to be completed within about 2-4 weeks. Please let me know if you require some more time.

Thank you so much for agreeing to review this submission!
