+++ paper/paper.bib
@@ -61,8 +61,8 @@
   title = {{redcapAPI}: Accessing data from REDCap projects using the API},
   author = {Benjamin Nutter and Stephen Lane},
   year = {2023},
-  doi = {10.5281/zenodo.11826},
-  note = {R package version 2.8.0},
+  doi = {10.5281/zenodo.10564836},
+  note = {R package version 2.8.3},
   url = {https://github.com/nutterb/redcapAPI/wiki},
 }
 
