### 2024-01-25 081f584
Update citation, update line 48 para w/ feedback
