+++ paper/paper.bib
@@ -59,11 +59,11 @@
 
 @Manual{redcapapi_cit,
   title = {{redcapAPI}: Accessing data from REDCap projects using the API},
-  author = {Benjamin Nutter and Stephen Lane},
-  year = {2023},
-  doi = {10.5281/zenodo.10564836},
+  author = {Shawn Garbett and Benjamin Nutter and Stephen Lane and Will Beasley and Jeffrey Horner and Jeremy Stephens and Marcus Lehr and Cole Beck and Savannah Obregon},
+  year = {2024},
+  doi = {10.5281/zenodo.10564837},
   note = {R package version 2.8.3},
-  url = {https://github.com/nutterb/redcapAPI/wiki},
+  url = {https://github.com/vubiostat/redcapAPI},
 }
 
 @Manual{redcapdm_cit,
