### 2024-01-26 #6277 @mikemahoney218
Seems like @rsh52 has been making changes in a branch here:
https://github.com/CHOP-CGTInformatics/REDCapTidieR/tree/joss-feedback-1

(I'm assuming the plan is to eventually merge this into main?)
