### 2024-01-29 #6277 @mikemahoney218
:wave: @rsh52 and @RhysPeploe , just wanted to share that I'm going to be OOO from February 2nd through the 9th (so, Friday through the end of next week). I'll be around _somewhat_ but will be much more delayed in responding on this issue; apologies in advance!

And as always, feel free to reach out if you have any questions or concerns :smile:
