### 2024-01-29 #6277 @rsh52
> Seems like @rsh52 has been making changes in a branch here: https://github.com/CHOP-CGTInformatics/REDCapTidieR/tree/joss-feedback-1
>
> (I'm assuming the plan is to eventually merge this into main?)

Just merged to main! Should I re-generate with the bot?
