### 2024-02-01 CHOP-CGTInformatics/REDCapTidieR#171 @RhysPeploe
[FEATURE] Remove unknown licence
openjournals/joss-reviews#6277

# Feature Request Description
Two licences are featured, however, all the information in the 'unknown'/non-MIT license is incorparated into the MIT license so it would be simplier to have just the one licence (the MIT one) and remove the other to avoid confusion. Unless there is a good reason to keep the two in parallel.

![image](https://github.com/CHOP-CGTInformatics/REDCapTidieR/assets/68282471/db01fa5a-ea68-4559-937f-8dac8faa7cf5)

![image](https://github.com/CHOP-CGTInformatics/REDCapTidieR/assets/68282471/b361c391-ec07-444f-92e0-7c0ae93bdece)

### Checklist

- [X] The issue is atomic
- [X] The issue description is documented
- [X] The issue title describes the problem succinctly
- [ ] Developers are assigned to the issue
- [X] Labels are assigned to the issue
