### 2024-02-01 CHOP-CGTInformatics/REDCapTidieR#171 @skadauke
Thanks Rhys. @rsh52 that seems easy enough to do!
