### 2024-02-01 CHOP-CGTInformatics/REDCapTidieR#171 @rsh52
Hi @RhysPeploe and @skadauke, thanks for pointing this out. Before I remove, I notice this is common among some other high-profile packages:

- [`gtsummary`](https://github.com/ddsjoberg/gtsummary)
- [`dplyr`](https://github.com/tidyverse/dplyr)
- [`ggplot2`](https://github.com/tidyverse/ggplot2)

From what I recall, this is due to a standard `LICENSE` file (not `LICENSE.md` where the MIT one exists) when you develop an R package in the standard framework. This was part of the initial commit from the start of the repository, and gets added to the `DESCRIPTION` file as `License: MIT + file LICENSE`.

Let me know your thoughts.
