### 2024-02-02 CHOP-CGTInformatics/REDCapTidieR#171 @rsh52
> Ah thank you for pointing this out, I hadn't noticed it elsewhere. Happy for you to keep it as is, I'll check this off the checklist

Great! Thank you. Closing this now then.
