### 2024-02-13 CHOP-CGTInformatics/REDCapTidieR#172 @RhysPeploe
[FEATURE] References Improvements
openjournals/joss-reviews/issues/6277

When checking the references on the paper, a few things came up:
- [x] line 108/109 the URL does not work, leads to a 404 error [REDCapDM](https://ubidi.github.io/REDCapDM/)
- [x] line 104 lowercase r for the programming language, suggest capitalisation
- [x] line 111 similar lowercase for redcap, suggest REDCap to be consistent
- [x] line 126 does not include the URL for the tibble package (although the link in the text line 64 works fine)
- [x] line 134 the url includes the " (website)" so that part needs unlinking from the hyperlink, as currently the url does not work

Once these are resolved my checklist for submission will be complete, thanks.

### Checklist

- [X] The issue is atomic
- [X] The issue description is documented
- [X] The issue title describes the problem succinctly
- [x] Developers are assigned to the issue
- [X] Labels are assigned to the issue
