### 2024-02-14 CHOP-CGTInformatics/REDCapTidieR#172 @rsh52
Hi @RhysPeploe. Thanks for the feedback. I looked at your recommendations and have a couple of questions:

> line 108/109 the URL does not work, leads to a 404 error [REDCapDM](https://ubidi.github.io/REDCapDM/)

It seems that team moved the URL, I will update it to the new link.

> line 104 lowercase r for the programming language, suggest capitalisation
> line 111 similar lowercase for redcap, suggest REDCap to be consistent

I do not see these on my end, is there any chance something is getting lost in translation?

> line 126 does not include the URL for the tibble package (although the link in the text line 64 works fine)

Can you confirm this link works: https://tibble.tidyverse.org/
This is what is hyperlinked to the tibble package, and works on my end.

> line 134 the url includes the " (website)" so that part needs unlinking from the hyperlink, as currently the url does not work

Apologies, it is difficult for me to follow these line numbers since they are different from [what is on GitHub](https://github.com/CHOP-CGTInformatics/REDCapTidieR/blob/main/paper/paper.md?plain=1). I'm wondering if the capitalization issues you're seeing might be solved there as well. Which link is this in reference to?
