+++ paper/paper.bib
@@ -71,7 +71,7 @@
   author = {João Carmezim and Judith Peñafiel and Pau Satorra and Esther García and Natàlia Pallarés and Cristian Tebé},
   year = {2023},
   note = {R package version 0.8.0},
-  url = {https://ubidi.github.io/REDCapDM/},
+  url = {https://bruigtp.github.io/REDCapDM/},
 }
 
 @Manual{tidyredcap_cit,
