### 2024-02-14 CHOP-CGTInformatics/REDCapTidieR#172 @RhysPeploe
![image](https://github.com/CHOP-CGTInformatics/REDCapTidieR/assets/68282471/0cea69db-33c3-48f8-b666-fb272c593a74)
