### 2024-02-14 CHOP-CGTInformatics/REDCapTidieR#172 @rsh52
Ah thank you @RhysPeploe

I see these are in the bib file. The first one is odd, is there a chance this is auto-enforced by the editorial bot? I have "R" capitalized in the file:

https://github.com/CHOP-CGTInformatics/REDCapTidieR/blob/78af9f881903ddd98e705bd6fe402357e13f0609/paper/paper.bib#L52-L54

For `redcapAPI` this is the actual [name of the package](https://cran.r-project.org/web/packages/redcapAPI/index.html), so IMO we should leave it in this format.

I can definitely fix the `tibble` URL citation. 👍
I will remove the "(website)" from the `skimr` citation 👍
