+++ paper/paper.bib
@@ -102,13 +102,14 @@
   author = {Elin Waring and Michael Quinn and Amelia McNamara and Eduardo {Arino de la Rubia} and Hao Zhu and Shannon Ellis},
   year = {2023},
   note = {R package version 2.1.5, https://github.com/ropensci/skimr/},
-  url = {https://docs.ropensci.org/skimr/ (website)},
+  url = {https://docs.ropensci.org/skimr/},
 }
 
 @Manual{tibble_cit,
   title = {tibble: Simple Data Frames},
   author = {Kirill Müller and Hadley Wickham},
   year = {2023},
+  url = {https://tibble.tidyverse.org/},
   note = {https://tibble.tidyverse.org/, https://github.com/tidyverse/tibble},
 }
 
