### 2024-02-14 CHOP-CGTInformatics/REDCapTidieR#172 @RhysPeploe
Try double parentheses {{title}}, this _should_ retain the way you have writen the title

@Manual{redcapr_cit,
  title = {{REDCapR: Interaction Between R and REDCap}},
  author = {Will Beasley},
  year = {2023},
  url = {https://ouhscbbmc.github.io/REDCapR/},
  note = {https://ouhscbbmc.github.io/REDCapR/, https://github.com/OuhscBbmc/REDCapR, https://www.ouhsc.edu/bbmc/, https://project-redcap.org},
}

My bad on the package name redcapAPI, keep that the same
