### 2024-02-14 56690b1
Double braces for REDCapR cit
