+++ paper/paper.bib
@@ -50,7 +50,7 @@
 }
 
 @Manual{redcapr_cit,
-  title = {REDCapR: Interaction Between R and REDCap},
+  title = {{REDCapR: Interaction Between R and REDCap}},
   author = {Will Beasley},
   year = {2023},
   url = {https://ouhscbbmc.github.io/REDCapR/},
