### 2024-02-14 #6277 @rsh52
@mikemahoney218 Would it be ok to update the current release tag (v1.0.0) and include updated notes? Or should we bump the package version to `1.0.0.9000` and make an all-new release?

For Zenodo, do I upload the source package after build (tar.gz file)?
