### 2024-02-14 #6277 @mikemahoney218
We just need a GitHub release that corresponds to this version of the software, no need to do a full release -- if you make a `v1.0.0-joss` tag to commemorate "this is the version of the software as accepted by JOSS", that's probably easiest?

For Zenodo, I _personally_ always have used [this tutorial which will create Zenodo archives every time you create a release](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content), because I like the idea of automatically archiving each release. If you do this, do it _before_ creating the new tag, and you'll need to make sure the title and authors (and ORCiDs) match the JOSS submission.

If you just want to create a single archive, then upload a zipped version of your source code -- not the built version. If it's helpful, a recent example for what this looks like: https://zenodo.org/records/10658194
