+++ pkgdown.yml
@@ -6,7 +6,7 @@ articles:
   diving_deeper: diving_deeper.html
   export_to_xlsx: export_to_xlsx.html
   glossary: glossary.html
-last_built: 2024-02-14T16:55Z
+last_built: 2024-02-14T17:07Z
 urls:
   reference: https://chop-cgtinformatics.github.io/REDCapTidieR/reference
   article: https://chop-cgtinformatics.github.io/REDCapTidieR/articles
