### 2024-02-14 #6277 @rsh52
@mikemahoney218 Thanks so much for helping clear this up. I believe everything should be available now:

- Zenodo Upload: https://zenodo.org/records/10658774
- REDCapTidieR JOSS Release: https://github.com/CHOP-CGTInformatics/REDCapTidieR/releases/tag/v1.0.0-joss
