### 2024-02-14 ec5c04d
Citations formatting, authors
