+++ paper/paper.bib
@@ -83,7 +83,7 @@
 }
 
 @Manual{labelled_cit,
-  title = {labelled: Manipulating Labelled Data},
+  title = {{labelled}: Manipulating Labelled Data},
   author = {Joseph Larmarange},
   year = {2023},
   note = {R package version 2.12.0},
@@ -91,14 +91,15 @@
 }
 
 @Manual{openxlsx2_cit,
-  title = {openxlsx2: Read, Write and Edit 'xlsx' Files},
+  title = {{openxlsx2}: Read, Write and Edit 'xlsx' Files},
+  author = {Jordan Mark Barbone and Jan Marvin Garbuszus},
   year = {2023},
   note = {R package version 1.1},
   url = {https://janmarvin.github.io/openxlsx2/},
 }
 
 @Manual{skimr_cit,
-  title = {skimr: Compact and Flexible Summaries of Data},
+  title = {{skimr}: Compact and Flexible Summaries of Data},
   author = {Elin Waring and Michael Quinn and Amelia McNamara and Eduardo {Arino de la Rubia} and Hao Zhu and Shannon Ellis},
   year = {2023},
   note = {R package version 2.1.5, https://github.com/ropensci/skimr/},
@@ -106,7 +107,7 @@
 }
 
 @Manual{tibble_cit,
-  title = {tibble: Simple Data Frames},
+  title = {{tibble}: Simple Data Frames},
   author = {Kirill Müller and Hadley Wickham},
   year = {2023},
   url = {https://tibble.tidyverse.org/},
@@ -114,9 +115,9 @@
 }
 
 @Misc{openssf_cit,
-  title = {OpenOpen Source Security Foundation_2023},
-  url={https://openssf.org/},
-  journal={Open Source Security Foundation},
+  title = {OpenSSF Best Practices Badge Program},
+  author = {Open Source Security Foundation},
+  url={https://www.bestpractices.dev/},
   publisher={The Linux Foundation},
   year={2023},
   month={Oct}
@@ -139,6 +140,6 @@
 @Manual{superheroes_cit,
   title = {Superhero Database},
   url = {https://www.superherodb.com/},
-  author = {Jeroen ter Lingen},
+  author = {Jeroen {ter Lingen}},
   year = {2023}
 }
