### 2024-02-14 #6277 @mikemahoney218
@rsh52 not sure if it's easier for you to change things or for me to open a PR -- think there's still an issue with how the OpenSSF citation is getting rendered (sorry, my fault). Think this would fix it:

```
@Misc{openssf_cit,
  title = {{OpenSSF Best Practices} Badge Program},
  author = {{Open Source Security Foundation}},
  url={https://www.bestpractices.dev/},
  publisher={The Linux Foundation},
  year={2023},
  month={Oct}
```

(new braces in the title and author fields)
