### 2024-02-14 ebd7984
Update paper.bib
Fix per JOSS review request
