+++ paper/paper.bib
@@ -115,8 +115,8 @@
 }
 
 @Misc{openssf_cit,
-  title = {OpenSSF Best Practices Badge Program},
-  author = {Open Source Security Foundation},
+  title = {{OpenSSF Best Practices} Badge Program},
+  author = {{Open Source Security Foundation}},
   url={https://www.bestpractices.dev/},
   publisher={The Linux Foundation},
   year={2023},
